fix function duplication in jit tests (#19583)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/19583

There are functions defined in "caffe2/test/cpp/jit/test_utils.h" for which Buck complains about duplication.  This splits those function definitions into a separate .cpp file.

Differential Revision: D15037723

fbshipit-source-id: e6c127b6e5d5b077a7c043e2e2c8ff0987ccab26
