+++ test/cpp/jit/CMakeLists.txt
@@ -2,7 +2,9 @@ set(JIT_TEST_ROOT ${TORCH_ROOT}/test/cpp/jit)
 
 add_executable(test_jit
   ${TORCH_ROOT}/test/cpp/common/main.cpp
-  ${JIT_TEST_ROOT}/test.cpp)
+  ${JIT_TEST_ROOT}/test.cpp
+  ${JIT_TEST_ROOT}/test_utils.cpp
+)
 
 target_link_libraries(test_jit PRIVATE torch gtest)
 target_compile_definitions(test_jit PRIVATE USE_GTEST)
